football_game/requirements.txt
pygame>=2.5.0
gymnasium>=0.29.0
numpy>=1.24.0
stable-baselines3>=2.0.0
torch>=2.0.0
tensorboard>=2.20.0
tqdm>=4.67.0
rich>=14.0.0
pyyaml>=6.0.0